Let 3.9.4 understand the changes that were included in 4.0.0, more importantly those around moving to long from int.
Remove client api changes.
Fix and remove broken tests.
Set LogRecord version to write only version 0
Change client version to 0.
Ignore metadata client tests.
Write log record v1 for tests writing eventnumbers greater than int.MaxValue.
Update check for chunk size to not run for chunk version 3.
Set record version for tests.
Change EventNumber.DeletedStream to int.MaxValue rather than long.MaxValue.
Update chunk version check for ongoing chunks.
Update MD5Hash verification.

+++ src/EventStore.Core/Data/StreamMetadata.cs
@@ -12,17 +12,17 @@ namespace EventStore.Core.Data
     {
         public static readonly StreamMetadata Empty = new StreamMetadata();
 
-        public readonly int? MaxCount;
+        public readonly long? MaxCount;
         public readonly TimeSpan? MaxAge;
         
-        public readonly int? TruncateBefore;
+        public readonly long? TruncateBefore;
         public readonly bool? TempStream;
 
         public readonly TimeSpan? CacheControl;
         public readonly StreamAcl Acl;
 
-        public StreamMetadata(int? maxCount = null, TimeSpan? maxAge = null,
-                              int? truncateBefore = null, bool? tempStream = null,
+        public StreamMetadata(long? maxCount = null, TimeSpan? maxAge = null,
+                              long? truncateBefore = null, bool? tempStream = null,
                               TimeSpan? cacheControl = null, StreamAcl acl = null)
         {
             if (maxCount <= 0)
@@ -73,9 +73,9 @@ namespace EventStore.Core.Data
             Check(reader.Read(), reader);
             Check(JsonToken.StartObject, reader);
 
-            int? maxCount = null;
+            long? maxCount = null;
             TimeSpan? maxAge = null;
-            int? truncateBefore = null;
+            long? truncateBefore = null;
             bool? tempStream = null;
             TimeSpan? cacheControl = null;
             StreamAcl acl = null;
@@ -93,7 +93,7 @@ namespace EventStore.Core.Data
                     {
                         Check(reader.Read(), reader);
                         Check(JsonToken.Integer, reader);
-                        maxCount = (int) (long) reader.Value;
+                        maxCount = (long) reader.Value;
                         break;
                     }
                     case SystemMetadata.MaxAge:
@@ -107,7 +107,7 @@ namespace EventStore.Core.Data
                     {
                         Check(reader.Read(), reader);
                         Check(JsonToken.Integer, reader);
-                        truncateBefore = (int) (long) reader.Value;
+                        truncateBefore = (long) reader.Value;
                         break;
                     }
                     case SystemMetadata.TempStream:
