+++ src/EventStore.Projections.Core.Tests/Services/position_tagging/multistream_position_tagger/when_updating_multistream_postion_tracker_from_a_tag.cs
@@ -18,7 +18,7 @@ namespace EventStore.Projections.Core.Tests.Services.position_tagging.multistrea
             var tagger = new MultiStreamPositionTagger(0, new []{"stream1", "stream2"});
             var tracker = new PositionTracker(tagger);
 
-            var newTag = CheckpointTag.FromStreamPositions(0, new Dictionary<string, int>{{"stream1", 1}, {"stream2", 2}});
+            var newTag = CheckpointTag.FromStreamPositions(0, new Dictionary<string, long>{{"stream1", 1}, {"stream2", 2}});
             tracker.UpdateByCheckpointTagInitial(newTag);
             _tag = tracker.LastTag;
             _tagger = new MultiStreamPositionTagger(0, new []{"stream1", "stream2"});
